\name{BSEHydroModels-package}
\alias{BSEHydroModels-package}
\alias{BSEHydroModels}
\docType{package}
\title{
\packageTitle{BSEHydroModels}
}
\description{
\packageDescription{BSEHydroModels}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{BSEHydroModels}
\packageIndices{BSEHydroModels}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{BSEHydroModels}

Maintainer: \packageMaintainer{BSEHydroModels}
}
\references{
~~ Literature or other references for background information ~~
}

\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
